+++ src/main/java/com/appraisaltool/model/UserGroupId.java
@@ -16,14 +16,14 @@ public class UserGroupId implements Serializable{
 		private static final long serialVersionUID = 3233124370476846883L;
 
 		@Column (insertable = false, updatable = false)
-		private Long userId;
+		private Integer userId;
 		
 		@Column (insertable = false, updatable = false)
-	    private Long groupId;
+	    private Integer groupId;
 	 
 	    // default constructor
 	 
-	    public UserGroupId(Long userId, Long groupId) {
+	    public UserGroupId(Integer userId, Integer groupId) {
 	        this.userId = userId;
 	        this.groupId = groupId;
 	    }
