+++ src/main/java/com/appraisaltool/repository/UserGroupRepository.java
@@ -12,23 +12,23 @@ import org.springframework.stereotype.Repository;
 import com.appraisaltool.model.UserGroup;
 
 @Repository
-public interface UserGroupRepository extends JpaRepository<UserGroup, Long>{
+public interface UserGroupRepository extends JpaRepository<UserGroup, Integer>{
 
 	@Query("SELECT u.groupId FROM UserGroup u WHERE u.userId = ?1")
-	public List<Long> findByUserGroupByUserId(Long userId);
+	public List<Integer> findByUserGroupByUserId(Integer userId);
 
 	@Transactional
 	@Modifying
 	@Query("DELETE FROM UserGroup u WHERE u.userId = ?1")
-	public void deleteAllByUserId(Long userId);
+	public void deleteAllByUserId(Integer userId);
 	
 	@Query("SELECT  eg.userId FROM UserGroup eg WHERE  eg.groupId  IN "
 			+ "(SELECT eg.groupId  FROM UserGroup eg INNER JOIN  User e ON eg.userId = e.userId "
 			+ "WHERE eg.userId = ?1) AND eg.userId != ?1")
-	public List<Long> findGroupPartners(Long userId);
+	public List<Integer> findGroupPartners(Integer userId);
 
 	@Query("SELECT u.groupId FROM UserGroup u WHERE u.userId = ?1")
-	public List<Long> getGroupsByUserId(Long userId);
+	public List<Integer> getGroupsByUserId(Integer userId);
 	
 	
 }
