+++ src/main/java/com/appraisaltool/service/AppraisalService.java
@@ -4,7 +4,7 @@ import com.appraisaltool.model.Appraisal;
 
 public interface AppraisalService {
 
-	public Appraisal getAppraisalById(Long id);
+	public Appraisal getAppraisalById(Integer id);
 	
 	public Appraisal createOrUpdateAppraisal(Appraisal id);
